Create New Patient button opens new empty row. Next steps: read patientList into columns, change columns into the ones I need.
