+++ src/App.js
@@ -1,40 +1,28 @@
 import './App.css';
-import { useEffect, useState } from 'react'
 import { Auth } from './components/auth.js'
-import { PatientsTableWithProviders } from './components/table.js'
-import { db } from './firebase-config.js'
-import { getDocs, collection } from 'firebase/firestore'
+import { PatientsTable } from './components/table.js'
+// import { db } from './firebase-config.js'
+// import { getDocs, collection } from 'firebase/firestore'
+import {
+  QueryClient,
+  QueryClientProvider,
+} from '@tanstack/react-query'
+
+
+const queryClient = new QueryClient()
 
 function App() {
 
-  const [patientList, setPatientList] = useState([])
-
-  const patientsCollectionRef = collection(db, "patients")
-
-  useEffect(() => {
-    const getPatientsList = async () => {
-      try {
-        const patientData = await getDocs(patientsCollectionRef)
-
-        const filteredPatientData = patientData.docs.map((doc) => ({...doc.data(), id: doc.id}))
-
-        setPatientList(filteredPatientData)
-      } catch (error) {
-        // handle error
-      }
-    }
-
-    getPatientsList()
-  }, [])
-
   return (
-    <div className="App">
-      <header className="App-header">
-        <Auth />
-        <PatientsTableWithProviders patients={patientList}/>
-      </header>
-    </div>
-  );
+    <QueryClientProvider client={queryClient}>
+      <div className="App">
+        <header className="App-header">
+          <Auth />
+          <PatientsTable />
+        </header>
+      </div>
+    </QueryClientProvider>
+  )
 }
 
-export default App;
+export default App
