+++ src/components/table.js
@@ -1,37 +1,51 @@
-import { useMemo, useState } from 'react';
-import { MantineReactTable, useMantineReactTable } from 'mantine-react-table';
+import { useMemo, useState, useEffect } from 'react';
+import { MantineReactTable, useMantineReactTable, createRow } from 'mantine-react-table';
 import { ActionIcon, Button, Flex, Text, Tooltip, Modal } from '@mantine/core';
 import { IconTrash } from '@tabler/icons-react';
-import {
-  QueryClient,
-  QueryClientProvider,
-  useMutation,
-  useQuery,
-  useQueryClient,
-} from '@tanstack/react-query';
+import { useCreatePatient, useGetPatients, useUpdatePatients, useDeletePatient } from '../helpers/hooks'
+import { validateRequired, validatePatient } from '../helpers/validation'
+import { db } from '../firebase-config.js'
+import { getDocs, collection } from 'firebase/firestore'
 
-const PatientsTable = (patients) => {
-  const [validationErrors, setValidationErrors] = useState({});
-  const [editedPatients, setEditedPatients] = useState({});
 
-  //call CREATE hook
-  const { mutateAsync: createPatient, isLoading: isCreatingPatient } =
-    useCreatePatient();
-  //call READ hook
+export const PatientsTable = () => {
+  const [validationErrors, setValidationErrors] = useState({})
+  const [patientList, setPatientList] = useState([])
+  const [editedPatients, setEditedPatients] = useState({})
+  const patientsCollectionRef = collection(db, "patients")
+
+
+  useEffect(() => {
+    const getPatientsList = async () => {
+    try {
+        const patientData = await getDocs(patientsCollectionRef)
+
+        const filteredPatientData = patientData.docs.map((doc) => ({...doc.data(), id: doc.id}))
+
+        setPatientList(filteredPatientData)
+    } catch (error) {
+        // handle error
+    }
+    }
+
+    getPatientsList()
+  }, [])
+
+  //what is this syntax:
+  const { mutateAsync: createPatient, isLoading: isCreatingPatient } = useCreatePatient()
+
   const {
     data: fetchedPatients = [],
     isError: isLoadingPatientsError,
     isFetching: isFetchingPatients,
     isLoading: isLoadingPatients,
   } = useGetPatients();
-  //call UPDATE hook
-  const { mutateAsync: updatePatients, isLoading: isUpdatingPatient } =
-    useUpdatePatients();
-  //call DELETE hook
-  const { mutateAsync: deletePatient, isLoading: isDeletingPatient } =
-    useDeletePatient();
 
-  //CREATE action
+  const { mutateAsync: updatePatients, isLoading: isUpdatingPatient } = useUpdatePatients();
+
+  const { mutateAsync: deletePatient, isLoading: isDeletingPatient } = useDeletePatient();
+
+  //handlers:
   const handleCreatePatient = async ({ values, exitCreatingMode }) => {
     const newValidationErrors = validatePatient(values);
     if (Object.values(newValidationErrors).some((error) => !!error)) {
@@ -43,14 +57,12 @@ const PatientsTable = (patients) => {
     exitCreatingMode();
   };
 
-  //UPDATE action
   const handleSavePatients = async () => {
     if (Object.values(validationErrors).some((error) => !!error)) return;
     await updatePatients(Object.values(editedPatients));
     setEditedPatients({});
   };
 
-  //DELETE action
 //   const openDeleteConfirmModal = (row) =>
 //     // modals.openConfirmModal({
 //     //   title: 'Are you sure you want to delete this Patient?',
@@ -65,6 +77,7 @@ const PatientsTable = (patients) => {
 //     //   onConfirm: () => deletePatient(row.original.id),
 //     });
 
+// CREATE THE TABLE
   const columns = useMemo(
     () => [
       {
@@ -114,31 +127,11 @@ const PatientsTable = (patients) => {
         }),
       },
       {
-        accessorKey: 'email',
-        header: 'Email',
-        mantineEditTextInputProps: ({ cell, row }) => ({
-          type: 'email',
-          required: true,
-          error: validationErrors?.[cell.id],
-          //store edited Patient in state to be saved later
-          onBlur: (event) => {
-            const validationError = !validateEmail(event.currentTarget.value)
-              ? 'Invalid Email'
-              : undefined;
-            setValidationErrors({
-              ...validationErrors,
-              [cell.id]: validationError,
-            });
-            setEditedPatients({ ...editedPatients, [row.id]: row.original });
-          },
-        }),
-      },
-      {
         accessorKey: 'state',
         header: 'State',
         editVariant: 'select',
         mantineEditSelectProps: ({ row }) => ({
-        //   data: usStates,
+          data: patientList,
           //store edited Patient in state to be saved later
           onChange: (value) =>
             setEditedPatients({
@@ -154,8 +147,8 @@ const PatientsTable = (patients) => {
   const table = useMantineReactTable({
     columns,
     data: fetchedPatients,
-    createDisplayMode: 'row', // ('modal', and 'custom' are also available)
-    editDisplayMode: 'cell', // ('modal', 'row', 'cell', and 'custom' are also available)
+    createDisplayMode: 'row', 
+    editDisplayMode: 'cell',
     enableEditing: true,
     enableRowActions: true,
     positionActionsColumn: 'last',
@@ -175,9 +168,9 @@ const PatientsTable = (patients) => {
     onCreatingRowSave: handleCreatePatient,
     renderRowActions: ({ row }) => (
       <Tooltip label="Delete">
-        {/* <ActionIcon color="red" onClick={() => openDeleteConfirmModal(row)}>
+        <ActionIcon color="red">
           <IconTrash />
-        </ActionIcon> */}
+        </ActionIcon>
       </Tooltip>
     ),
     renderBottomToolbarCustomActions: () => (
@@ -201,15 +194,8 @@ const PatientsTable = (patients) => {
     renderTopToolbarCustomActions: ({ table }) => (
       <Button
         onClick={() => {
-          table.setCreatingRow(true); //simplest way to open the create row modal with no default values
-          //or you can pass in a row object to set default values with the `createRow` helper function
-          // table.setCreatingRow(
-          //   createRow(table, {
-          //     //optionally pass in default values for the new row, useful for nested data or other complex scenarios
-          //   }),
-          // );
-        }}
-      >
+          table.setCreatingRow(true || false)
+        }}>
         Create New Patient
       </Button>
     ),
@@ -224,111 +210,3 @@ const PatientsTable = (patients) => {
   return <MantineReactTable table={table} />;
 };
 
-//CREATE hook (post new Patient to api)
-function useCreatePatient() {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: async (Patient) => {
-      //send api update request here
-      await new Promise((resolve) => setTimeout(resolve, 1000)); //fake api call
-      return Promise.resolve();
-    },
-    //client side optimistic update
-    onMutate: (newPatientInfo) => {
-      queryClient.setQueryData(['Patients'], (prevPatients) => [
-        ...prevPatients,
-        {
-          ...newPatientInfo,
-          id: (Math.random() + 1).toString(36).substring(7),
-        },
-      ]);
-    },
-    // onSettled: () => queryClient.invalidateQueries({ queryKey: ['Patients'] }), //refetch Patients after mutation, disabled for demo
-  });
-}
-
-//READ hook (get Patients from api)
-function useGetPatients(patients) {
-  return useQuery({
-    queryKey: ['Patients'],
-    queryFn: async () => {
-      //send api request here
-      await new Promise((resolve) => setTimeout(resolve, 1000)); //fake api call
-      return Promise.resolve(patients);
-    },
-    refetchOnWindowFocus: false,
-  });
-}
-
-//UPDATE hook (put Patients in api)
-function useUpdatePatients() {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: async (Patients) => {
-      //send api update request here
-      await new Promise((resolve) => setTimeout(resolve, 1000)); //fake api call
-      return Promise.resolve();
-    },
-    //client side optimistic update
-    onMutate: (newPatients) => {
-      queryClient.setQueryData(['Patients'], (prevPatients) =>
-        prevPatients?.map((Patient) => {
-          const newPatient = newPatients.find((u) => u.id === Patient.id);
-          return newPatient ? newPatient : Patient;
-        }),
-      );
-    },
-    // onSettled: () => queryClient.invalidateQueries({ queryKey: ['Patients'] }), //refetch Patients after mutation, disabled for demo
-  });
-}
-
-//DELETE hook (delete Patient in api)
-function useDeletePatient() {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: async (PatientId) => {
-      //send api update request here
-      await new Promise((resolve) => setTimeout(resolve, 1000)); //fake api call
-      return Promise.resolve();
-    },
-    //client side optimistic update
-    onMutate: (PatientId) => {
-      queryClient.setQueryData(['Patients'], (prevPatients) =>
-        prevPatients?.filter((Patient) => Patient.id !== PatientId),
-      );
-    },
-    // onSettled: () => queryClient.invalidateQueries({ queryKey: ['Patients'] }), //refetch Patients after mutation, disabled for demo
-  });
-}
-
-const queryClient = new QueryClient();
-
-const PatientsTableWithProviders = () => (
-  //Put this with your other react-query providers near root of your app
-  <QueryClientProvider client={queryClient}>
-    {/* <ModalsProvider> */}
-      <PatientsTable />
-    {/* </ModalsProvider> */}
-  </QueryClientProvider>
-);
-
-export {PatientsTableWithProviders};
-
-const validateRequired = (value) => !!value?.length;
-const validateEmail = (email) =>
-  !!email.length &&
-  email
-    .toLowerCase()
-    .match(
-      /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/,
-    );
-
-function validatePatient(Patient) {
-  return {
-    firstName: !validateRequired(Patient.firstName)
-      ? 'First Name is Required'
-      : '',
-    lastName: !validateRequired(Patient.lastName) ? 'Last Name is Required' : '',
-    email: !validateEmail(Patient.email) ? 'Incorrect Email Format' : '',
-  };
-}
